Initial RenderEngine2 with getting render data in separate thread.
